refactor(fwGdcmIO): const-ness/keywords/logs
- Fix const-ness
- Use MessageDialog in SSurfaceSgmentationWriter instead of LOGS
- Simplify for( ... ) loops
- Update keywords: throw/override
- Update license date
- Remove dead code

+++ SrcLib/io/fwGdcmIO/src/fwGdcmIO/helper/Fiducial.cpp
@@ -98,9 +98,9 @@ bool Fiducial::contains3DDistances(const SPTR(::fwMedData::SeriesDB)& seriesDB)
                         {
                             const ::fwData::Point::sptr point1 = *pointList->getCRefPoints().begin();
                             const ::fwData::Point::sptr point2 = *(++pointList->getCRefPoints().begin());
-                            const int frameNumber1             =
+                            const size_t frameNumber1          =
                                 ::fwGdcmIO::helper::DicomDataTools::convertPointToFrameNumber(image, point1);
-                            const int frameNumber2 =
+                            const size_t frameNumber2 =
                                 ::fwGdcmIO::helper::DicomDataTools::convertPointToFrameNumber(image, point2);
                             if(frameNumber1 != frameNumber2)
                             {
